Graph_RAG/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.0
langchain==0.1.0
langchain-google-genai==0.0.10
neo4j==5.13.0
streamlit==1.28.1
PyPDF2==3.0.1
